+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/ui/InterfaceIO.kt
@@ -1,14 +1,8 @@
 package world.gregs.voidps.engine.client.ui
 
 import world.gregs.voidps.engine.client.ui.detail.InterfaceComponentDetail
-import world.gregs.voidps.engine.client.ui.detail.InterfaceDetail
 
 interface InterfaceIO {
-    fun sendOpen(inter: InterfaceDetail)
-    fun sendClose(inter: InterfaceDetail)
-    fun notifyClosed(inter: InterfaceDetail)
-    fun notifyOpened(inter: InterfaceDetail)
-    fun notifyRefreshed(inter: InterfaceDetail)
     fun sendPlayerHead(component: InterfaceComponentDetail)
     fun sendAnimation(component: InterfaceComponentDetail, animation: Int)
     fun sendNPCHead(component: InterfaceComponentDetail, npc: Int)
